Move boards.j5.arduino to boards.arduino.uno
