+++ j5/backends/hardware/sb/arduino.py
@@ -9,7 +9,6 @@ from serial.tools.list_ports_common import ListPortInfo
 from j5.backends import CommunicationError
 from j5.backends.hardware.env import NotSupportedByHardwareError
 from j5.backends.hardware.j5.arduino import ArduinoHardwareBackend
-from j5.boards.j5.arduino import FIRST_ANALOGUE_PIN
 from j5.boards.sb.arduino import SBArduinoBoard
 from j5.components import GPIOPinMode
 from j5.components.derived import UltrasoundInterface
@@ -109,7 +108,7 @@ class SBArduinoHardwareBackend(
 
     def _update_digital_pin(self, identifier: int) -> None:
         """Write the stored value of a digital pin to the Arduino."""
-        if identifier >= FIRST_ANALOGUE_PIN:
+        if identifier >= SBArduinoBoard.FIRST_ANALOGUE_PIN:
             raise RuntimeError("Reached an unreachable statement.")
         pin = self._digital_pins[identifier]
         char: str
@@ -141,7 +140,7 @@ class SBArduinoHardwareBackend(
 
     def _read_analogue_pin(self, identifier: int) -> float:
         """Read the value of an analogue pin from the Arduino."""
-        if identifier >= FIRST_ANALOGUE_PIN + 4:
+        if identifier >= SBArduinoBoard.FIRST_ANALOGUE_PIN + 4:
             raise NotSupportedByHardwareError(
                 "Arduino Uno firmware only supports analogue pins 0-3 (IDs 14-17)",
             )
@@ -204,11 +203,11 @@ class SBArduinoHardwareBackend(
             echo_pin_identifier: int,
     ) -> None:
         """Verify the validity of a pair of ultrasound pins."""
-        if trigger_pin_identifier >= FIRST_ANALOGUE_PIN:
+        if trigger_pin_identifier >= SBArduinoBoard.FIRST_ANALOGUE_PIN:
             raise NotSupportedByHardwareError(
                 "Ultrasound functions not supported on analogue pins",
             )
-        if echo_pin_identifier >= FIRST_ANALOGUE_PIN:
+        if echo_pin_identifier >= SBArduinoBoard.FIRST_ANALOGUE_PIN:
             raise NotSupportedByHardwareError(
                 "Ultrasound functions not supported on analogue pins",
             )
