+++ j5/boards/arduino/__init__.py
@@ -1,0 +1,7 @@
+"""Arduino Board Definitions."""
+
+from .uno import ArduinoUno
+
+__all__ = [
+    "ArduinoUno",
+]
