--- j5/boards/j5/__init__.py
@@ -1,9 +1,0 @@
-"""Base classes for generalisations of boards that J5 supports."""
-
-from .arduino import AnaloguePin, ArduinoUno, PinNumber
-
-__all__ = [
-    'ArduinoUno',
-    'AnaloguePin',
-    'PinNumber',
-]
