+++ tests/boards/sb/test_arduino.py
@@ -6,7 +6,6 @@ from typing import TYPE_CHECKING, Optional, Set
 import pytest
 
 from j5.backends import Backend
-from j5.boards.j5 import AnaloguePin
 from j5.boards.sb import SBArduinoBoard
 from j5.components import GPIOPin, GPIOPinInterface, GPIOPinMode, LEDInterface
 from j5.components.derived import UltrasoundInterface, UltrasoundSensor
@@ -138,7 +137,7 @@ def test_uno_pins() -> None:
     for i in range(2, 14):
         assert type(uno.pins[i]) is GPIOPin
 
-    for j in AnaloguePin:
+    for j in SBArduinoBoard.AnaloguePin:
         assert type(uno.pins[j]) is GPIOPin
